image-poc/pyproject.toml
[project]
name = "cardnews-cli"
version = "0.1.0"
description = "A CLI tool to automatically generate Instagram card news."
authors = [{ name = "Gemini", email = "gemini@google.com" }]
dependencies = [
    "click",
    "Pillow",
    "PyYAML",
    "google-generativeai",
]
requires-python = ">=3.9"

[project.scripts]
cardnews = "cardnews.cli:main"
